FinalProject/requirements.txt
Flask==2.0.3
Flask-Cors==3.0.10
gunicorn==20.1.0; python_version > '3.0'
gunicorn==19.10.0; python_version < '3.0'
google-cloud-vision==2.7.2
google-cloud-storage==2.0.0; python_version < '3.7'
google-cloud-storage==2.1.0; python_version > '3.6'
google-cloud-datastore==2.4.0
sendgrid==6.9.7